Rename WithSet
